third_party/blink/web_tests/http/tests/websocket/send-expected.txt

WebSocket send test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected.
PASS data is FIRST_MESSAGE_TO_SEND
PASS data is SECOND_MESSAGE_TO_SEND
Closed.
PASS successfullyParsed is true

TEST COMPLETE
